+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0 - 2022-02-05
+### Added
+- Configuration of listeners using attributes (`\Tjovaisas\Bundle\DelayedEventBundle\Attribute\AsDelayedEventListener`)
+- Library will try to guess method's name based on the passed event when configuring listeners
+### Removed
+- Support for Symfony ^4.0
+- Support for Symfony >=5.0 <5.3
+
 ## 0.3.0 - 2022-01-30
 ### Added
 - Support for Symfony ^6.0